+++ markdown/CHANGELOG.md
@@ -2,8 +2,10 @@ Changelog WebOpal
 ==================================
 
 ##???: WebOpal v0.3 ##
+- NEW FEATURE: Bugreport / Featurevorschläge
 - NEW FEATURE: Optimierung der Ladezeiten
 - NEW FEATURE: Dynamisches Hinzufügen und Löschen von Strukturen
+- NEW FEATURE: Hin- und Herspringen von Editor zu Editor mit Alt-WASD
 - Minor: Drücken von Enter in der Befehlszeile wird den Code ausführen
 - Minor: Drücken von Strg+Enter führt den Code von überall aus.
 - Minor: Logo hinzugefügt